+++ pedidos/src/main/java/com/lanchonete/pedidos/repository/PedidosRepository.java
@@ -1,8 +1,9 @@
 package com.lanchonete.pedidos.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import com.lanchonete.pedidos.domain.Pedidos;
 
-public interface PedidosRepository extends JpaRepository<Pedidos, int[]>{
+public interface PedidosRepository extends JpaRepository<Pedidos, Long[]>{
   
 }
